requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
black==24.4.1
celery==5.3.4
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.40.0
colorama==0.4.6
convoy-python==0.2.0
cryptography==41.0.4
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.6
django-allauth==0.54.0
django-cors-headers==4.3.1
django-encrypted-model-fields==0.6.5
django-environ==0.11.2
django-prometheus==2.3.1
django-redis==5.4.0
django-rest-auth==0.9.5
django-taggit==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fastjsonschema==2.18.1
filelock==3.12.3
flower==2.0.1
humanize==4.8.0
idna==3.6
kombu==5.3.2
Markdown==3.5.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
pipenv==2023.9.7
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.2.1
vine==5.0.0
virtualenv==20.24.4
wcwidth==0.2.8
